--- facenet-retinaface-pytorch/face_dataset/README.md
@@ -1,1 +1,0 @@
-数据库图片
