--- facenet-retinaface-pytorch/img/README.md
@@ -1,1 +1,0 @@
-此处存放图像保存文件
