--- facenet-retinaface-pytorch/model_data/README.md
@@ -1,1 +1,0 @@
-此处存放权重文件
