--- facenet-retinaface-pytorch/nets/facenet.py
@@ -1,86 +1,0 @@
-import torch.nn as nn
-from torch.nn import functional as F
-
-from nets.inception_resnetv1 import InceptionResnetV1
-from nets.mobilenet import MobileNetV1
-
-
-class mobilenet(nn.Module):
-    def __init__(self):
-        super(mobilenet, self).__init__()
-        self.model = MobileNetV1()
-        del self.model.fc
-        del self.model.avg
-
-    def forward(self, x):
-        # 160,160,3 -> 20,20,256
-        x = self.model.stage1(x)
-        # 20,20,256 -> 10,10,512
-        x = self.model.stage2(x)
-        # 10,10,512 -> 5,5,1024
-        x = self.model.stage3(x)
-        return x
-
-class inception_resnet(nn.Module):
-    def __init__(self):
-        super(inception_resnet, self).__init__()
-        self.model = InceptionResnetV1()
-
-    def forward(self, x):
-        x = self.model.conv2d_1a(x)
-        x = self.model.conv2d_2a(x)
-        x = self.model.conv2d_2b(x)
-        x = self.model.maxpool_3a(x)
-        x = self.model.conv2d_3b(x)
-        x = self.model.conv2d_4a(x)
-        x = self.model.conv2d_4b(x)
-        x = self.model.repeat_1(x)
-        x = self.model.mixed_6a(x)
-        x = self.model.repeat_2(x)
-        x = self.model.mixed_7a(x)
-        x = self.model.repeat_3(x)
-        x = self.model.block8(x)
-        return x
-        
-    
-class Facenet(nn.Module):
-    def __init__(self, backbone="mobilenet", dropout_keep_prob=0.5, embedding_size=128, num_classes=None, mode="train"):
-        super(Facenet, self).__init__()
-        if backbone == "mobilenet":
-            self.backbone = mobilenet()
-            flat_shape = 1024
-        elif backbone == "inception_resnetv1":
-            self.backbone = inception_resnet()
-            flat_shape = 1792
-        else:
-            raise ValueError('Unsupported backbone - `{}`, Use mobilenet, inception_resnetv1.'.format(backbone))
-        self.avg = nn.AdaptiveAvgPool2d((1,1))
-        self.Dropout = nn.Dropout(1 - dropout_keep_prob)
-        self.Bottleneck = nn.Linear(flat_shape, embedding_size,bias=False)
-        self.last_bn = nn.BatchNorm1d(embedding_size, eps=0.001, momentum=0.1, affine=True)
-        if mode == "train":
-            self.classifier = nn.Linear(embedding_size, num_classes)
-
-    def forward(self, x):
-        x = self.backbone(x)
-        x = self.avg(x)
-        x = x.view(x.size(0), -1)
-        x = self.Dropout(x)
-        x = self.Bottleneck(x)
-        x = self.last_bn(x)
-        x = F.normalize(x, p=2, dim=1)
-        return x
-
-    def forward_feature(self, x):
-        x = self.backbone(x)
-        x = self.avg(x)
-        x = x.view(x.size(0), -1)
-        x = self.Dropout(x)
-        x = self.Bottleneck(x)
-        before_normalize = self.last_bn(x)
-        x = F.normalize(before_normalize, p=2, dim=1)
-        return before_normalize, x
-
-    def forward_classifier(self, x):
-        x = self.classifier(x)
-        return x
