--- facenet-retinaface-pytorch/static/css/index.css
@@ -1,67 +1,0 @@
-/* 布局样式 */
-body {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  font-family: sans-serif;
-}
-
-#video {
-  max-width: 100%;
-  margin-bottom: 20px;
-  border: 5px solid #ddd;
-  box-shadow: 1px 1px 4px rgba(0, 0, 0, 0.3);
-  background-color: #333;
-}
-
-#photo {
-  width: 450px;height: 450px;
-  margin-bottom: 20px;
-  border: 5px solid #ddd;
-  box-shadow: 1px 1px 4px rgba(0, 0, 0, 0.3);
-  background-color: aliceblue;
-}
-
-#canvas {
-  margin-top: 20px;
-  border: 5px solid #ddd;
-  box-shadow: 1px 1px 4px rgba(0, 0, 0, 0.3);
-}
-
-button {
-  align-content: center;
-  padding: 10px 20px;
-  font-size: 1rem;
-  background-color: #007bff;
-  color: #fff;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  margin: 0 auto;
-  transition: all 0.2s ease-in-out;
-}
-.inline-div {
-    margin-top: 20px;
-    width: 600px;
-    height: 200px;
-    text-align: center;
-}
-
-button:hover {
-  background-color: #0069d9;
-}
-.scale{
-    width: 450px;height: 450px;
-    background-color: aliceblue;
-}
-img {
-    width: 450px;height: 450px;
-    object-fit: scale-down;
-    background-color: aliceblue;
-}
-/* 隐藏样式 */
-.hidden {
-  display: none;
-}
