--- facenet-retinaface-pytorch/templates/index.html
@@ -1,105 +1,0 @@
-<!doctype html>
-<html lang="">
-<head>
-  <title>人脸识别</title>
-  <meta charset="utf-8">
-  <link rel='stylesheet' type='text/css' href='../static/css/index.css'>
-  <script src="../static/js/jquery.min.js"></script>
-</head>
-<body>
-  <h1>人脸识别</h1>
-  <div class="scale"><img id='photo' src="../img/img.png" alt=""></div>
-  <!--  accept=“image/*” 指只接收图片，不限jpg和gif-->
-  <input id="up-file" type="file" class="hidden" accept="image/*">
-
-  <!-- 功能按钮 -->
-  <div class="inline-div">
-      <button id="snapshot-button">图片录入</button>
-      <button id="analyze-button" class="hidden">图片分析</button>
-      <button id="save-button" class="hidden">图片保存</button>
-
-      <button id="start-button">摄像头识别</button>
-      <button id="stop-button" class="hidden">停止摄像头</button>
-  </div>
-
-  <script type="text/javascript">
-    var socket = io.connect();
-    $(document).ready(function (){
-        socket.on('connect', function() {
-          console.log('Connected to server');
-        });
-
-        socket.on('message', function(data) {
-            $('#photo').attr('src', 'data:image/jpeg;base64,' + data);
-        });
-
-        $('#start-button').click(function(){
-          socket.emit('start_camera', {});
-        });
-
-        // 关闭流媒体
-        $('#stop-button').click(function() {
-          socket.emit('stop_stream',{})
-        });
-    })
-
-  </script>
-  <script>
-
-    /** 点击按钮，上传新图片 **/
-    $(document).on("click", "#snapshot-button", function () {
-        // 点击图片的同时，点击上传文件的input
-        $('#up-file').click();
-        // 将上传的图片显示到页面上
-        $(document).on("change", "#up-file", function () {
-            var fileObj = $("#up-file")[0];
-            var reader = new FileReader();
-            // 转换成功后的操作，img.src即为转换后的DataURL
-            reader.onload = function(e) {
-                if (reader.readyState === 2) { //加载完毕后赋值
-                    photo.src = e.target.result;
-                    console.log(img.src);
-                }
-            }
-            reader.readAsDataURL(fileObj.files[0]);
-
-            $('#analyze-button').show();
-            $('#snapshot-button').hide();
-        });
-    });
-
-    // 上传图片到后台
-    $(document).on('click','#analyze-button',function () {
-        var imgData = $('#photo').attr('src');
-        // 创建FormData对象
-        var formData = new FormData();
-        formData.append("image", dataURItoBlob(imgData));
-        $.ajax({
-          url: '/upload',
-          type: 'POST',
-          data: formData,
-          contentType: false,
-          processData: false,
-            datatype:'text',
-          success: function(data) {
-            console.log("Success: " + data);
-          },
-          error: function(xhr, status, error) {
-            console.log("Error: " + error);
-          }
-        });
-    })
-
-    // 将Base64编码的数据转换成Blob对象
-    function dataURItoBlob(dataURI) {
-      var byteString = atob(dataURI.split(',')[1]);
-      var ab = new ArrayBuffer(byteString.length);
-      var ia = new Uint8Array(ab);
-      for (var i = 0; i < byteString.length; i++) {
-        ia[i] = byteString.charCodeAt(i);
-      }
-      return new Blob([ab], { type: 'image/jpeg' });
-    }
-  </script>
-</body>
-</html>
