--- facenet-retinaface-pytorch/templates/l-m-r.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-    <style>
-        *{
-            margin: 0;
-            padding: 0;
-        }
-        h1{
-            text-align: center;
-        }
-        .main{
-            margin: 20px;
-            width: 100%;
-            height: 500px;
-        }
-        .left{
-            float: left;
-            width: 200px;
-            height: 100%;
-            background-color: #3898b1;
-        }
-        .right{
-            float: right;
-            width: 200px;
-            height: 100%;
-            background-color: #ce7486;
-        }
-        .middle{
-            margin: 0 200px;
-            height: 100%;
-            background-color: #cbaf91;
-        }
-    </style>
-</head>
-<body>
-<h1>人脸识别</h1>
-<div class="main">
-    <div class="left">
-        <button id="snapshot-button">图片录入</button>
-        <button id="analyze-button" class="hidden">图片分析</button>
-        <button id="save-button" class="hidden">图片保存</button>
-    </div>
-    <div class="right">
-
-    </div>
-    <div class="middle">
-
-    </div>
-</div>
-
-</body>
-</html>
