--- facenet-retinaface-pytorch/templates/t.html
@@ -1,13 +1,0 @@
-<html>
-  <head>
-    <title>Video Streaming Demonstration</title>
-  </head>
-  <body>
-    <h1>Video Streaming Demonstration</h1>
-
-    <img src="{{ url_for('video_feed') }}"  height="500" alt="" id="t-img">
-  </body>
-<script>
-
-</script>
-</html>
