--- facenet-retinaface-pytorch/templates/tmp.html
@@ -1,186 +1,0 @@
-<!DOCTYPE html>
-<html lang="">
-<head>
-	<title>人脸识别</title>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<style>
-		*{
-            margin: 0 auto;
-            padding: 0;
-        }
-		h1{
-			text-align: center;
-			margin: 20px;
-		}
-		.main-div{
-			width: 100%;
-			height: 90%;
-			text-align: center;
-			vertical-align: center;
-		}
-		.left-div{
-			float: left;
-			background-color: aquamarine;
-			width: 200px;
-			height: 90%;
-			margin: 0 auto;
-		}
-		.right-div{
-			float: right;
-			background-color: aquamarine;
-			width: 200px;
-			height: 90%;
-		}
-		.center-div{
-			margin: 0 200px;
-			background-color: antiquewhite;
-			height: 90%;
-		}
-		button {
-			align-content: center;
-			padding: 10px 20px;
-			font-size: 1rem;
-			background-color: #007bff;
-			color: #fff;
-			border: none;
-			border-radius: 5px;
-			cursor: pointer;
-			transition: all 0.2s ease-in-out;
-			margin: 30px auto;
-		}
-		img {
-			width: 450px;height: 450px;
-			object-fit: scale-down;
-			background-color: aliceblue;
-		}
-		#photo {
-		  width: 450px;height: 450px;
-		  margin-bottom: 20px;
-		  border: 5px solid #ddd;
-		  box-shadow: 1px 1px 4px rgba(0, 0, 0, 0.3);
-		  background-color: aliceblue;
-		  padding: 40px;
-		}
-
-		/* 隐藏样式 */
-		.hidden {
-		  display: none;
-		}
-	</style>
-	<script src="../static/js/jquery.min.js"></script>
-	<script src="../static/js/socket.io.min.js"></script>
-</head>
-<body>
-	<h1>人脸识别</h1>
-	<div class="main-div">
-		<div class="left-div">
-			<button id="snapshot-button">图片录入</button>
-			<button id="analyze-button">图片分析</button>
-			<button id="save-button">图片保存</button>
-		</div>
-		<div class="right-div">
-			<button id="start-button">摄像头识别</button>
-			<button id="stop-button">停止摄像头</button>
-		</div>
-		<div class="center-div">
-			<div class="scale"><img id='photo' src="../img/img.png" alt=""></div>
-			<!--  accept=“image/*” 指只接收图片，不限jpg和gif-->
-			<input id="up-file" type="file" class="hidden" accept="image/*">
-		</div>
-	</div>
-
-</body>
-
-  <script>
-	  var socket = io.connect();
-    $(document).ready(function (){
-        socket.on('connect', function() {
-          console.log('Connected to server');
-        });
-
-        socket.on('message', function(data) {
-            $('#photo').attr('src', 'data:image/jpeg;base64,' + data);
-        });
-
-        $('#start-button').click(function(){
-          socket.emit('start_camera', {});
-        });
-
-        // 关闭流媒体
-        $('#stop-button').click(function() {
-          socket.emit('stop_stream',{})
-        });
-    })
-
-    /** 点击按钮，上传新图片 **/
-    $(document).on("click", "#snapshot-button", function () {
-        // 点击图片的同时，点击上传文件的input
-        $('#up-file').click();
-        // 将上传的图片显示到页面上
-        $(document).on("change", "#up-file", function () {
-            var fileObj = $("#up-file")[0];
-            var reader = new FileReader();
-            // 转换成功后的操作，img.src即为转换后的DataURL
-            reader.onload = function(e) {
-                if (reader.readyState === 2) { //加载完毕后赋值
-                    photo.src = e.target.result;
-                    console.log(img.src);
-                }
-            }
-
-			reader.readAsDataURL(fileObj.files[0]);
-        });
-    });
-
-    // 上传图片到后台
-    $(document).on('click','#analyze-button',function () {
-		var imgData = $('#photo').attr('src');
-		// 创建FormData对象
-		var formData = new FormData();
-		formData.append("image", dataURItoBlob(imgData));
-		$.ajax({
-		  url: '/upload',
-		  type: 'POST',
-		  data: formData,
-			contentType: false,
-          processData: false,
-			datatype: 'text',
-		  success: function(data) {
-			  alert(data)
-			  $("#photo").attr("src", "data:image/jpeg;base64," + data);
-		  },
-		  error: function(xhr, status, error) {
-			console.log("Error: " + error);
-		  }
-		});
-    })
-
-	$(document).on('click','#save-button',function () {
-		var name=prompt("请输入图片名称");
-		if (name!=null){
-			$.ajax({
-				url:'/save',
-				type:'POST',
-				data: {name:name},
-				success:function (data){
-					alert('save success')
-				}
-			});
-		}
-
-
-	})
-
-    // 将Base64编码的数据转换成Blob对象
-    function dataURItoBlob(dataURI) {
-      var byteString = atob(dataURI.split(',')[1]);
-      var ab = new ArrayBuffer(byteString.length);
-      var ia = new Uint8Array(ab);
-      for (var i = 0; i < byteString.length; i++) {
-        ia[i] = byteString.charCodeAt(i);
-      }
-      return new Blob([ab], { type: 'image/jpeg' });
-    }
-  </script>
-</html>
