--- facenet-retinaface-pytorch/util/interface.py
@@ -1,3 +1,0 @@
-# -*- coding:utf-8 -*-
-# author:peng
-# Date：2023/4/26 20:21
